support for parsing URDF model descriptions in DART
Link and Joints are extracted recursively from the root link that is automatically
determined or can be provided by the user. Joints are stores as frames in the
DART model, Links define the geometric representation. The package path is
determined automatically and a optional mesh file extension can be provided
to load meshes from other file types than provided in the URDF description.